+++ django_kafka/__init__.py
@@ -18,7 +18,7 @@ if TYPE_CHECKING:
 
 logger = logging.getLogger(__name__)
 
-__version__ = "0.5.1"
+__version__ = "0.5.2"
 
 __all__ = [
     "autodiscover",